Remove seechanges.c file, fix some compilation troubles
